+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on http://keepachangelog.com/en/1.0.0/
 and this project adheres to http://semver.org/spec/v2.0.0.html.
 
+## [unreleased]
+
+- Add method ColorCodes.ByName
+
 ## [0.2.0] 2022-03-20
 
 - Add types ColorCodes and AttributeCodes